fix a bug in clearDataBase

+++ HW3/dbsys-hw3/clearDataBase.py
@@ -21,7 +21,8 @@ for file in fileNames:
   
 db = Database(dataDir='./data');
 
-tmpRel = db.relations();
+tmpRel = list( db.relations() );
 for relKey in tmpRel:
-  db.removeRelation( relKey );
+  if relKey.startswith('tmp'):
+    db.removeRelation( relKey );
